Use ElelectronCryptoService to handle desktop-specific tasks
