+++ common/src/abstractions/platformUtils.service.ts
@@ -4,7 +4,6 @@ export abstract class PlatformUtilsService {
     identityClientId: string;
     getDevice: () => DeviceType;
     getDeviceString: () => string;
-    isDesktopClient: boolean;
     isFirefox: () => boolean;
     isChrome: () => boolean;
     isEdge: () => boolean;
