+++ electron/src/services/electronPlatformUtils.service.ts
@@ -20,7 +20,6 @@ import { ElectronConstants } from '../electronConstants';
 
 export class ElectronPlatformUtilsService implements PlatformUtilsService {
     identityClientId: string;
-    isDesktopClient = true;
 
     private deviceCache: DeviceType = null;
